% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_George.R
\name{ST_George}
\alias{ST_George}
\title{Sant Georges is Disease-specific instrument}
\usage{
ST_George(dt)
}
\arguments{
\item{dt}{data.frame}
}
\value{
dt and SYMPTOM,ACTIVIT,IMPACTS,TOTALSG variables
}
\description{
More detailed description.
}
\examples{
BD <- data.frame(
"sg1" = c(0, 1, 1, 0),
"sg2" = c(0, 1, 1, 0),
"sg3" = c(0, 1, 1, 0),
"sg4" = c(0, 1, 1, 0),
"sg5" = c(0, 1, 1, 0),
"sg6" = c(0, 1, 1, 0),
"sg7" = c(0, 1, 1, 0),
"sg8" = c(0, 1, 1, 0),
"sg9" = c(0, 1, 1, 0),
"sg10" = c(0, 1, 1, 0),
"sg111" = c(0, 1, 1, 0),
"sg112" = c(0, 1, 1, 0),
"sg113" = c(0, 1, 1, 0),
"sg114" = c(0, 1, 1, 0),
"sg115" = c(1, 1, 1, 1),
"sg116" = c(1, 1, 1, 1),
"sg117" = c(1, 1, 1, 1),
"sg121" = c(1, 1, 1, 1),
"sg122" = c(1, 1, 1, 1),
"sg123" = c(1, 1, 1, 1),
"sg124" = c(1, 1, 1, 1),
"sg125" = c(1, 1, 1, 1),
"sg126" = c(1, 1, 1, 1),
"sg131" = c(1, 1, 1, 1),
"sg132" = c(1, 1, 1, 1),
"sg133" = c(1, 1, 1, 1),
"sg134" = c(1, 1, 1, 1),
"sg135" = c(1, 1, 1, 1),
"sg136" = c(1, 1, 1, 1),
"sg137" = c(1, 1, 1, 1),
"sg138" = c(1, 1, 1, 1),
"sg141" = c(1, 1, 1, 1),
"sg142" = c(1, 1, 1, 1),
"sg143" = c(1, 1, 1, 1),
"sg144" = c(1, 1, 1, 1),
"sg151" = c(1, 1, 1, 1),
"sg152" = c(1, 1, 1, 1),
"sg153" = c(1, 1, 1, 1),
"sg154" = c(1, 1, 1, 1),
"sg155" = c(1, 0, 1, 0),
"sg156" = c(1, 0, 1, 0),
"sg157" = c(1, 0, 1, 0),
"sg158" = c(1, 0, 1, 0),
"sg159" = c(1, 0, 1, 0),
"sg161" = c(1, 0, 1, 0),
"sg162" = c(1, 0, 1, 0),
"sg163" = c(1, 0, 1, 0),
"sg164" = c(1, 0, 1, 0),
"sg165" = c(1, 0, 1, 0),
"sg17" = c(1, 0, 1, 0)
)
ST_George(BD)
}
