% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StGeorge.R
\docType{package}
\name{StGeorge}
\alias{StGeorge}
\title{StGeorge Package}
\description{
Disease-specific instrument designed to measure impact on overall health in patients with obstructive airways disease.
}
\author{
Ray Puig \email{raypuig@gmail.com}
}
